pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["yijun"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
google-cloud-pubsub = "^2.21.1"
pydantic = "^2.7.0"
pydantic-settings = "^2.2.1"


[tool.poetry.group.dev.dependencies]
black = "^24.4.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
